[FIX] Community컨트롤러 테스트 로그인 코드 개선 및 결과값 타입 수정(long -> int) [FIX] PostDAO 매개변수 어노테이션 제거 [FIX] PostService 반환 타입 수정(long -> int) [FIX] PostServiceImpl 에러 처리 관련 코드 수정(6,8번 코멘트) [FIX] 파일명 파싱 관련 변수 제거 및 Getter 추가로 파일명 쪼개기 기능 추가

+++ src/main/java/com/frade/service/community/PostService.java
@@ -11,7 +11,7 @@ import com.frade.dto.community.PostDTO;
 public interface PostService {
 	
 	//작성한 게시글 db에 저장
-	public long savePost(PostDTO post, MultipartFile[] files);
+	public int savePost(PostDTO post, MultipartFile[] files);
 
 	//게시글 목록 데이터 가져오기
     // 1.페이징 수학 계산
